requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
pg8000==1.30.5
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
Flask-Migrate==4.0.7
cryptography==42.0.5
pyOpenSSL==24.0.0
endesive==2.0.10
Flask-Mail==0.10.0
twilio==9.0.0
reportlab==4.0.4
pytest
pytest-flask
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Flask-Limiter==3.5.0
structlog==24.1.0
Flask-Caching==2.1.0
